Move functions into the corresponding class
